requirements.txt
discord.py[voice]==2.3.2
python-vlc==3.0.20123
python-dotenv==1.0.0
PyNaCl==1.5.0
audioop-lts==0.1
tmdbsimple==2.9.1
requests==2.31.0  # Required by tmdbsimple
